chore: address flaky test failures related to global partitioning (#3369)
